+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -------------------
 - Fixed customer birthday and title
 - Fixed product purchasePrice push and pull
+- Fixed temp directory and added fallback
 
 2.0.14
 ------